refactor: update backend logic
