+++ daytistics/supabase/functions/add-to-roadmap/deno.json
@@ -8,6 +8,7 @@
     "@shared/validation": "../../_shared/validation.ts",
     "@application/conversations": "../../_application/conversations.ts",
     "@application/tokens_budgets": "../../_application/tokens_budgets.ts",
-    "@application/daytistics": "../../_application/daytistics.ts"
+    "@application/daytistics": "../../_application/daytistics.ts",
+    "@application/roadmap": "../../_application/roadmap.ts"
   }
 }
