+++ daytistics/supabase/functions/add-to-roadmap/index.ts
@@ -10,10 +10,9 @@ import { initSentry, initSupabase } from "@shared/adapters";
 import { validateZodSchema } from "@shared/validation";
 import { z } from "npm:zod";
 import { User } from "https://jsr.io/@supabase/supabase-js/2.48.1/src/index.ts";
+import { addToRoadmap } from "@application/roadmap";
 
 const AddToRoadmapObject = z.object({
-  // union of features and bugs
-  roadmap: z.union([z.literal("features"), z.literal("bugs")]),
   title: z.string(),
   description: z.string(),
 });
@@ -35,7 +34,7 @@ Deno.serve(async (req) => {
     if (validateError) return validateError;
 
     try {
-      await updateRoadmap(body.roadmap, body.title, body.description, user!);
+      await addToRoadmap(body.title, body.description);
     } catch (error) {
       Sentry.captureException(error as Error);
       await Sentry.flush();
@@ -63,147 +62,3 @@ Deno.serve(async (req) => {
     });
   }
 });
-
-async function updateRoadmap(
-  roadmap: "features" | "bugs",
-  title: string,
-  description: string,
-  user: User,
-) {
-  const projectNodeIdResponse = await fetch(`https://api.github.com/graphql`, {
-    method: "POST",
-    headers,
-    body: JSON.stringify({
-      query:
-        `query{user(login: "leo-gall") {projectV2(number: ${PROJECT_ID}){id}}}`,
-    }),
-  });
-
-  const projectNodeId = (await projectNodeIdResponse.json()).data.user.projectV2
-    .id;
-
-  const statusFieldIdResponse = await fetch(
-    `https://api.github.com/graphql`,
-    {
-      method: "POST",
-      headers,
-      body: JSON.stringify({
-        query: `query {
-          node(id: "${projectNodeId}") {
-            ... on ProjectV2 {
-              fields(first: 20) {
-                nodes {
-                  ... on ProjectV2Field {
-                    id
-                    name
-                  }
-                  ... on ProjectV2IterationField {
-                    id
-                    name
-                    configuration {
-                      iterations {
-                        startDate
-                        id
-                      }
-                    }
-                  }
-                  ... on ProjectV2SingleSelectField {
-                    id
-                    name
-                    options {
-                      id
-                      name
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }`,
-      }),
-    },
-  );
-
-  const statusFieldData = (await statusFieldIdResponse.json()).data.node.fields
-    .nodes
-    .find((field: { name: string }) => field.name === "Status")!;
-
-  const draftBody = description +
-    "\n\n---\n\nThis item was automatically created from the app by " +
-    user.id;
-
-  const statusOptionId = statusFieldData.options.find(
-    (option: { name: string }) =>
-      option.name ===
-        (roadmap === "features" ? "Todo (Features)" : "Todo (Bugs)"),
-  )?.id;
-
-  if (!statusOptionId) {
-    throw new Error("Failed to find the appropriate status option ID.");
-  }
-
-  const addItemToProjectResponse = await fetch(
-    `https://api.github.com/graphql`,
-    {
-      method: "POST",
-      headers,
-      body: JSON.stringify({
-        query: `mutation {
-          addProjectV2DraftIssue(input: {
-            projectId: "${projectNodeId}",
-            title: "[FROM-APP] ${title}",
-            body: """${draftBody}""",
-          }) {
-            projectItem {
-              id
-            }
-          }
-        }`,
-      }),
-    },
-  );
-
-  if (!addItemToProjectResponse.ok) {
-    throw new Error(
-      `Failed to add draft issue to project: ${await addItemToProjectResponse
-        .text()}`,
-    );
-  }
-
-  const { data: { addProjectV2DraftIssue: { projectItem } } } =
-    await addItemToProjectResponse.json();
-
-  const updateStatusResponse = await fetch(
-    `https://api.github.com/graphql`,
-    {
-      method: "POST",
-      headers,
-      body: JSON.stringify({
-        query: `mutation {
-          updateProjectV2ItemFieldValue(input: {
-            projectId: "${projectNodeId}",
-            itemId: "${projectItem.id}",
-            fieldId: "${statusFieldData.id}",
-            value: { singleSelectOptionId: "${statusOptionId}" }
-          }) {
-            projectV2Item {
-              id
-            }
-          }
-        }`,
-      }),
-    },
-  );
-
-  if (!updateStatusResponse.ok) {
-    throw new Error(
-      `Failed to update status field: ${await updateStatusResponse.text()}`,
-    );
-  }
-
-  if (!addItemToProjectResponse.ok) {
-    throw new Error(
-      `Failed to add item to project: ${await addItemToProjectResponse.text()}`,
-    );
-  }
-}
